@@ -5,14 +5,14 @@ version: 0.0.6
 homepage: https://github.com/kfman/flutter.exchange_room_calendar
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  xml: ^5.3.1
-  http: ^0.13.4
+  xml: ^6.2.2
+  http: ^0.13.5
 
 dev_dependencies:
   flutter_test:
