@@ -2,10 +2,13 @@ import 'dart:convert';
 
 /// Credentials to access the Exchange server
 class ConnectionCredentials {
-  final String username;
-  final String password;
-  final String domain;
-  final String serverUrl;
+  final String? username;
+  final String? password;
+  final String? domain;
+  final String? serverUrl;
+  final String appId;
+  final String tenantId;
+
 
   String get authorization =>
       'Basic ' + base64Encode(utf8.encode('$username@$domain:$password'));
@@ -13,6 +16,6 @@ class ConnectionCredentials {
   /// [username] with access to the rooms calendar and its [password]
   /// using the [domain] e.d. __mydomain.local__ on the server
   /// with the [serverUrl] ending on __/EWS/Exchange.asmx__
-  ConnectionCredentials(
+  ConnectionCredentials(this.appId, this.tenantId,
       {this.username, this.password, this.domain, this.serverUrl});
 }
