@@ -2,19 +2,19 @@ import 'package:xml/xml.dart';
 
 /// [Room] or resource on an Exchange server
 class Room {
-  String name;
-  String emailAddress;
-  String routingType;
-  String mailboxType;
+  String? name;
+  String? emailAddress;
+  String? routingType;
+  String? mailboxType;
 
   Room();
 
   /// Extracts an Room from a XML [element]
   Room.fromXml(XmlElement element) {
-    var innerElement = element.getElement('t:Id');
-    name = innerElement.getElement('t:Name').text;
-    emailAddress = innerElement.getElement('t:EmailAddress').text;
-    routingType = innerElement.getElement('t:RoutingType').text;
-    mailboxType = innerElement.getElement('t:MailboxType').text;
+    var innerElement = element.getElement('t:Id')!;
+    name = innerElement.getElement('t:Name')!.text;
+    emailAddress = innerElement.getElement('t:EmailAddress')!.text;
+    routingType = innerElement.getElement('t:RoutingType')!.text;
+    mailboxType = innerElement.getElement('t:MailboxType')!.text;
   }
 }
